Replaced old-style ValidParams class
